Some polish
I did a bit of (unnecessary) polish
Added a tree-spawner
Set wagon rotation to trigger on the rotation of the lever, centered the player character on the wagon (this makes it harder to grab the handles).
Introduced new bug: Some of the wagons reset the position when reset. I have chosen this doesn't ruin the experience